Do not enqueue procedure from user global marked as no-decompiled
Restore fix done in #416
